% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copss.R
\name{copss_dp_ErrMat}
\alias{copss_dp_ErrMat}
\title{COPSS method with DP algorithm}
\usage{
copss_dp_ErrMat(X, W, Kmax = 15)
}
\arguments{
\item{X}{N*p matrix
W if normalized is TRUE, then W is inverse of estimated_sigma_square(X); 
  if is FALSE, then W = rep(1, ncol(X)).}

\item{W}{if normalized is TRUE, then W is inverse of estimated_sigma_square(X); 
if is FALSE, then W = rep(1, ncol(X)).}

\item{Kmax}{an upper bound on the true number of change-points}
}
\value{
A list of two matrices mat_e and mat_o. The sum of CV errors.
  mat_o: which uses 'odd' data to estimate, CV error matrix on 'even' data.
  mat_e: which uses 'even' data to estimate, CV error matrix on 'odd' data.
}
\description{
COPSS method with DP algorithm to find the optimal number of change-points
}
\details{
copss_dp_ErrMat is the same as copss_dp.
}
