% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cp_hdmean_dp}
\alias{cp_hdmean_dp}
\title{The Dynamic Programming Method}
\usage{
cp_hdmean_dp(X, W, Kmax)
}
\arguments{
\item{X}{a n*p data matrix}

\item{W}{the normalized scale of the data}

\item{Kmax}{scalar, the maximum number of changepoints}
}
\description{
Change-point detection algorithm for the high dimension mean change model
using the dynamic programming.
}
\details{
Dynamic Programming of changepoint detection in order to get all 
possible changepoints given a maximum number of changepoints, \code{Kmax}.
}
\examples{
n <- 100
tau2n <- c(0.1, 0.13, 0.15, 0.23, 0.25, 0.40, 0.44, 0.65, 0.76, 0.78, 0.81)
tau2n <- c(0.1, 0.40, 0.65, 0.8)
(tau_true <- floor(tau2n * n))
x <- gen_piecewise_const_vector(n, tau_true)
X <- matrix(x, n, 1)
# W <- estimated_sigma_square(X)
W <- rep(1, ncol(X))
Kmax <- length(tau_true)
tau_mat <- cp_hdmean_dp(X, W, Kmax)
tau_mat[Kmax, ]
}
