% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cp_hdmean_op}
\alias{cp_hdmean_op}
\title{The Optimal Partitioning Method of High Dimension Mean Change Model}
\usage{
cp_hdmean_op(X, W, pen)
}
\arguments{
\item{X}{a n*p data matrix}

\item{W}{the normalized scale of the data}

\item{pen}{the penalty of the cost function}
}
\value{
the locations of estimated change-points
}
\description{
Optimal Partitioning method of changepoint detection.
}
\details{
Change-point detection algorithm for the high dimension mean change model
using Optimal Partitioning.
}
\examples{
n <- 100
tau2n <- c(0.1, 0.13, 0.15, 0.23, 0.25, 0.40, 0.44, 0.65, 0.76, 0.78, 0.81)
tau2n <- c(0.1, 0.40, 0.65, 0.8)
(tau_true <- floor(tau2n * n))
x <- gen_piecewise_const_vector(n, tau_true)
X <- matrix(x, n, 1)
W <- estimated_sigma_square(X)
# W <- rep(1, ncol(X))
pen <- log(n)^1.1
op_list <- cp_hdmean_op(X, W, pen)
(op_list$cptau)

}
