% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagenerating.R
\name{cpb_true_tau}
\alias{cpb_true_tau}
\title{Change-point model B -- generate the true change-point locations}
\usage{
cpb_true_tau(n)
}
\arguments{
\item{n}{sample size}
}
\value{
the locations of the change-points, without including 0 and n.
}
\description{
Change-point model B -- generate the true change-point locations
}
\examples{
  tau <- cpb_true_tau(2048)
  n <- 400
  tau_true <- cpb_true_tau(n)
  tau_vec <- c(0, sort(tau_true), n)
  min(diff(tau_vec))
}
