% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{criterion_cp_num}
\alias{criterion_cp_num}
\title{The distribution of change-point num}
\usage{
criterion_cp_num(K_mat, Ktrue)
}
\arguments{
\item{K_mat}{Matrix, the estimated change-point matrix, the results of OP, PELT, DP.
A n_sim * n_algorithms matrix, whose element is the value of K_hat by cv.}

\item{Ktrue}{Vector, the true change-point vector}
}
\value{
A matrix, crit_mat
}
\description{
The criterion of the distribution of change-point num
}
