% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{criterion_hausdorff}
\alias{criterion_hausdorff}
\title{Criterion of Hausdorff distance}
\usage{
criterion_hausdorff(hausdorff_OE, hausdorff_UE)
}
\arguments{
\item{hausdorff_OE}{vector, overestimated}

\item{hausdorff_UE}{vector, underestimated}
}
\value{
A matrix, crit_mat
}
\description{
Calculate the criterion of Hausdorff distance between vectors.
}
