% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cv_objfun}
\alias{cv_objfun}
\title{\eqn{C(\tau^{train}; Xtest)}}
\usage{
cv_objfun(Xtrain, Xtest, W, tau)
}
\arguments{
\item{Xtrain, }{matrix with odd index, which is used to estimate tau.}

\item{Xtest, }{matrix with even index, used to construct the prediction error.}

\item{W, }{matrix with p*p, as the global normalizer, by dividide W[j]}

\item{tau, }{the vector of change-points, including 0, and n. 
Usually tau is obtained by Xtrian, and this cv_objfun is used to 
calculate the cross-validation squared loss.}
}
\value{
mean of the cross validation error
}
\description{
The function of definition of \eqn{C(\tau^{train}; Xtest)}
}
\details{
\deqn{C(\tau^{train}; Xtest) = \sum_{l=0}^{L}
  \sum_{i=\tau_{l}+1}^{\tau{l+1}}
  (Xtrain_i - \bar(Xtest)_{\tau_{l}, \tau{l+1}})^{2} / W }
}
