% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{CVMat_optimal_Khat}
\alias{CVMat_optimal_Khat}
\title{Function converts errMat to optimal Khat result}
\usage{
CVMat_optimal_Khat(res)
}
\arguments{
\item{res}{A list of mat_o and mat_e, both are error matrices.}
}
\value{
A list of opt_cploc_e, opt_cploc_o, opt_skou, opt_Khat.
  opt_cploc_o, opt_cploc_e, are both arrays, the optimal location;
  opt_skou, (sick_or_not, K_hat, K_diff, OE, UE)
  opt_Khat, a list of integers, the optimal change-point number, 
  from (O, E, inter, union) data.
}
\description{
Function converts errMat to optimal Khat result
}
