% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copss.R
\name{dpCVErrMat}
\alias{dpCVErrMat}
\title{Calculate the cv error matrix}
\usage{
dpCVErrMat(Xtrain, Xtest, W, tauMat)
}
\arguments{
\item{Xtrain}{train data used to estimate the change-points}

\item{Xtest}{test data used to calculate the cv error}

\item{W}{vector with p-length as the global normalizer.}

\item{tauMat}{candidate tau's matrix estimated by DP based on data Xtrain}
}
\value{
errMat with columns of 
  c("objfit", "objfun", "K_hat", "K_diff", "OE", "UE"),
  then followed with columns of \eqn{\tau_{1}} to \eqn{\tau_{Kmax}}
}
\description{
Calculate the cv error matrix from the tauMat
}
