% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copss.R
\name{dpFitErrMat}
\alias{dpFitErrMat}
\title{Calculate the fit error matrix with DP algorithm}
\usage{
dpFitErrMat(X, W, tauMat)
}
\arguments{
\item{X}{N*p matrix}

\item{W}{vector with p-length as the global normalizer.}

\item{tauMat}{candidate tau's matrix estimated by DP based on data Xtrain}
}
\value{
An error matrix, with c("objfit", "K_hat") 
  and c("tau_1",..., "tau_Kmax") as columns.
}
\description{
Calculate the fit error matrix with DP algorithm without hausdorff distance
}
