% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copsshausdorff.R
\name{dpTauMat2FitErrMat}
\alias{dpTauMat2FitErrMat}
\title{Calculate the fit error matrix using tau matrix obtained by DP algorithm}
\usage{
dpTauMat2FitErrMat(X, W, tau2n, tauMat)
}
\arguments{
\item{X}{N*p matrix}

\item{W}{vector with p-length as the global normalizer.}

\item{tau2n}{the true tau/n ratio.}

\item{tauMat}{candidate tau's matrix estimated by DP based on data Xtrain}
}
\value{
An error matrix, with c("objfit", "K_hat", "K_diff", "OE", "UE") 
  and "tau_1", "tau_Kmax" as columns.
}
\description{
Calculate the fit error matrix using tau matrix obtained by DP algorithm
}
\details{
Calculate the fit error matrix using tau matrix obtained by DP algorithm
and it includes the hausdorff distance. 
The function is often used in BIC methods.
}
