% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{errMat_optimal_Khat}
\alias{errMat_optimal_Khat}
\title{Optimal change-point number using high-dimension BIC}
\usage{
errMat_optimal_Khat(errMat)
}
\arguments{
\item{errMat}{The error matrix}
}
\value{
A list of optimal location and optimal number.
  opt_cploc, array, the optimal location;
  opt_Khat, an integer, the optimal change-point number.
}
\description{
Function gives the optimal change-point number using High-dimension BIC
}
