% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagenerating.R
\name{gen_ARMA_residual}
\alias{gen_ARMA_residual}
\title{Generate ARMA residual}
\usage{
gen_ARMA_residual(n, p, b1 = 0, b2 = 0, residualform = "normal")
}
\arguments{
\item{n}{integer, sample size}

\item{p}{integer, dimension}

\item{b1}{\eqn{\beta_1}}

\item{b2}{\eqn{\beta_2}}

\item{residualform}{string, the residual form, which can take value of 
"normal", "t", "chisq", "uniform", 
or "quasinormal" but only when p = 1.
All the variance of these residual data are normalized to 1.
So we can only focus on different shapes of different distrubution.}
}
\value{
A random matrix n*p
}
\description{
Generate ARMA residual data
}
\details{
The mean-change model: 
\deqn{X_i = \mu_{k} + \Sigma^{1/2} * z_i}

Generating error term 
\eqn{\epsilon_{i} = \Sigma^{1/2} * z_i} from an ARMA model,
which specifies the correlations between the j-1 and the j-th dimensions, 
i.e. correlations between \eqn{\epsilon_{i, j-1}} and \eqn{\epsilon_{i, j}}.
\deqn{\epsilon_{ij} - \beta_{1} \epsilon_{i, j-1} = e_{ij} + \beta_{2} e_{i,j-1}}
where \eqn{e_{ij}} and \eqn{e{i, j-1}} are the white noise with zero mean.

Four correlation structures:
\itemize{
  \item{(1) \eqn{\beta_1, \beta_{2} = 0, 0}   the independent(IND) case;}
  \item{(2) \eqn{\beta_1, \beta_{2} = 3/4, 0}   the autoregressive(AR) case;}
  \item{(3) \eqn{\beta_1, \beta_{2} = 0, 3/4}   the moving-average(MA) case;}
  \item{(4) \eqn{\beta_1, \beta_{2} = 1/2, 1/2}   the ARMA case;}
 }
}
