% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagenerating.R
\name{gen_data_IND_mcp}
\alias{gen_data_IND_mcp}
\title{Generate independent data in mcp model}
\usage{
gen_data_IND_mcp(
  n,
  p,
  r,
  umax,
  tau2n,
  residualform = "normal",
  diff_signal = FALSE,
  sig_const = 1,
  amplify_signal = 1,
  verbose = FALSE
)
}
\arguments{
\item{n}{Sample size}

\item{p}{Dimension}

\item{r}{Sparsity ratio. There are p*r dims that have change-points.
Set r=1 when it is in low dimension, i.e. p is small.}

\item{umax}{the constant (or the max value) of each element of
the generated vector}

\item{tau2n}{The locations of change-points relative to the sample size,
without 0 and 1 included, just the pure change-points.}

\item{residualform}{string, the residual form, which can take value of 
"normal", "t", "chisq", "uniform", 
or "quasinormal" but only when p = 1.
All the variance of these residual data are normalized to 1.
So we can only focus on different shapes of different distrubution.}

\item{diff_signal}{logical.
Allow different signals along different dimensions or not.}

\item{sig_const}{scalar, default 1, the constant before the residual.}

\item{amplify_signal}{scalar, default 1, the constant before the signal.}

\item{verbose}{bool, default FALSE, whether to return a list of 
data with (X, mu, eps) or X.}
}
\value{
X data matrix n*p
}
\description{
Generate data, independent case, multiple change-point model (mean change)
}
\details{
IND case: {X_i = u_k + sig_const * eps_i}. 
  X_i is an obsevation vector with p-dim. 
  Independent case has 0 correlation between components of eps_{i}, 
  namely 0 correlation between eps_{ij} and eps_{i,j-1}.
}
\examples{
n <- 200; p <- 500; r <- 0.05; umax <- 0.9
tau2n <- c(0.2, 0.40, 0.60, 0.80)
residualform="normal"; diff_signal=FALSE; sig_const=1;  amplify_signal=1
set.seed(12345)
X <- gen_data_IND_mcp(n, p, r, umax, tau2n, residualform="normal", 
  diff_signal=FALSE, sig_const=1, amplify_signal=1)
}
