% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagenerating.R
\name{gen_hcv_ARMA_data}
\alias{gen_hcv_ARMA_data}
\title{Generate high-dimension CV ARMA data}
\usage{
gen_hcv_ARMA_data(
  n,
  p,
  r,
  umax,
  tau2n,
  cor_structure = "IND",
  residualform = "normal",
  diff_signal = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{n}{integer, sample size}

\item{p}{integer, dimension}

\item{r}{the ratio where floor(p*r) dims in total have a change.}

\item{umax}{maximum element in the \eqn{\mu = (0, a, 0, a, \cdots)}}

\item{tau2n}{The locations of change-points relative to the sample size,
without 0 and 1 included, just the pure change-points.}

\item{cor_structure}{"IND", "AR", "MA", "ARMA"}

\item{residualform}{string, the residual form, which can take value of 
"normal", "t", "chisq", "uniform", 
or "quasinormal" but only when p = 1.
All the variance of these residual data are normalized to 1.
So we can only focus on different shapes of different distrubution.}

\item{diff_signal}{logical.
Allow different signals along different dimensions or not.
There are three options: TRUE, FALSE, "permutation", which is corresponding}

\item{verbose}{bool, default FALSE, whether to return a list of 
data with (X, mu, eps) or X.}
}
\value{
X data matrix n*p
}
\description{
Generate ARMA data for high-dimension CV
}
\details{
Assume diff_signal = FALSE forever; and amplify_signal=1 forever
}
\examples{
  tau2n <- c(0.2, 0.4, 0.6, 0.8)
  n <- 40; p <- 100; r <- 1; umax <- 1;
  dt <- gen_hcv_ARMA_data(
      n, p, r, umax, tau2n, cor_structure="IND", 
      residualform="normal", 
      diff_signal=TRUE, verbose=TRUE)
  Xu <- dt$Xu
}
