% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagenerating.R
\name{gen_IND_residual}
\alias{gen_IND_residual}
\title{Generate IND residual}
\usage{
gen_IND_residual(n, p, residualform = "normal")
}
\arguments{
\item{n}{integer, sample size}

\item{p}{integer, dimension}

\item{residualform}{string, the residual form, which can take value of 
"normal", "t", "chisq", "uniform", 
or "quasinormal" but only when p = 1.
All the variance of these residual data are normalized to 1.
So we can only focus on different shapes of different distrubution.}
}
\value{
A random matrix of n*p.
}
\description{
Generate independent residual data
}
