% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagenerating.R
\name{gen_MultNorm_residual}
\alias{gen_MultNorm_residual}
\title{Generate residual from a mutivariate normal distribution}
\usage{
gen_MultNorm_residual(n, p, rho = 1)
}
\arguments{
\item{n}{sample size}

\item{p}{dimension}

\item{rho}{\eqn{S_{p*p} = \rho^{|i-j|}} the element in the covariance matrix,
default \eqn{\rho=1}}
}
\value{
n*p residual matrix
}
\description{
Simulate from a zero-mean Multivariate Normal Distribution
}
\details{
Generate residual data from a Multivariate normal distribution.
}
