% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagenerating.R
\name{gen_piecewise_const_vector}
\alias{gen_piecewise_const_vector}
\title{Generate piecewise constant vector}
\usage{
gen_piecewise_const_vector(n, tau_true, umax = 1)
}
\arguments{
\item{n}{the length of the generated vector, same as the sample size}

\item{tau_true}{true change-points, \eqn{\tau}'s not the tau2n's, 
without the 0 nor n included.}

\item{umax}{the max value of each element in the \eqn{\theta}-vector.
\eqn{\theta} is further repeated to generate the result vector. 
In other words, umax is the max value of the generated vecor.}
}
\value{
a n-length vetor
}
\description{
Generate piecewise constant vector in 1D with n-length
}
\details{
Generate piecewise constant data with tau_true as the true change-points.

Generate piecewise constant vector with length n, 
which starts with 0, such as [0, a, a, a, 0, 0,...]. 
The number of the repeated a's or repeated 0's are determine by 
\eqn{\tau_{j+1} - \tau_{j}}.
}
\examples{
n <- 200
tau_true <- c(50, 100)
x <- gen_piecewise_const_vector(n, tau_true, umax=3)
}
