% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagenerating.R
\name{genconst_mat_signal}
\alias{genconst_mat_signal}
\title{Generate (Kn+1)*p true signal matrix \eqn{\mu}}
\usage{
genconst_mat_signal(Kn, umax, p, r, diff_signal)
}
\arguments{
\item{Kn}{the number of change-points}

\item{umax}{the constant (or the max value) of each element of
the generated vector}

\item{p}{dimension}

\item{r}{the ratio where floor(p*r) dims in total have a change.}

\item{diff_signal}{logical.
Allow different signals along different dimensions or not.
There are three options: TRUE, FALSE, "permutation", which is corresponding}
}
\value{
a (Kn+1)*p matrix, with each row as a mean vector in each
  segmentation, for example, row-vector generated by 
  \code{\link{genconst_Kn_theta}}
}
\description{
Generate (Kn+1)*p true signal matrix \eqn{\mu} in \deqn{X = \mu + \epsilon}
}
\details{
This function generate (Kn+1)*p true signal matrix \eqn{\mu} by extending 
theta_true[i] to a p-len vector with \eqn{nr = p * r} components being 
theta_true[i] (make sure nr is at least one).
}
