% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{hausdorff_dist_vector}
\alias{hausdorff_dist_vector}
\title{Hausdorff distance between vectors}
\usage{
hausdorff_dist_vector(B, eB, nX)
}
\arguments{
\item{B}{Vector, B_{m} the true change-points locations}

\item{eB}{Vector, eB_{n} the estimated change-points locations.
For example \code{eB = c(20, 40, 60)}, without 0 and nX included.}

\item{nX}{Interger, the length of X matrix}
}
\value{
A list of OE (over estimated) and UE(under estimated)
}
\description{
Calculate the hausdorff distance between vectors
}
