% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copss.R
\name{hBIC_dp}
\alias{hBIC_dp}
\title{High-dimensional BIC (hBIC) method with DP algorithm}
\usage{
hBIC_dp(X, W, Kmax, pen)
}
\arguments{
\item{X}{N*p matrix}

\item{W}{if normalized is TRUE, then W is inverse of estimated_sigma_square(X); 
if is FALSE, then W = rep(1, ncol(X)).}

\item{Kmax}{an upper bound on the true number of change-points}

\item{pen}{The penalty}
}
\value{
The error matrix, with columns of "objfit", "objfun", "K_hat", 
  followed with the locations of change-points.
}
\description{
The hBIC method with DP algorithm to find the optimal number of change-points
}
\details{
The hBIC criterion is calculated with different values of \eqn{\alpha}'s 
  in finding the optimal number of change-points. 
  
The classic BIC criterion is defined as follows:
\deqn{BIC(L) = loss_function + L {\log(n)}^{1 + \alpha}}
  where loss_function = cv_objfun(X, X, W, cp_hat), L is the estimated muber 
  of change-points.
  
The hBIC criterion is defined as follows:
\deqn{hBIC(L) = loss_function + L (\xi_{n, p} + p)}
  Then the penalty in hBIC can be formulated as:
\deqn{penalty = p + \xi_{n, p} 
  = p + c_0 \times \sqrt{trace_R2} \times \log(n)^{1 + \alpha} }
}
