% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copss.R
\name{hCV_dp}
\alias{hCV_dp}
\title{High-dimensional cross-validation (hCV) method with DP algorithm}
\usage{
hCV_dp(X, W, Kmax = 15)
}
\arguments{
\item{X}{N*p matrix}

\item{W}{if normalized is TRUE, then W is inverse of estimated_sigma_square(X); 
if is FALSE, then W = rep(1, ncol(X)).}

\item{Kmax}{an upper bound on the true number of change-points}
}
\value{
A list of two matrices mat_e and mat_o. 
  mat_o: which uses 'odd' data to estimate, CV error matrix on 'even' data.
  mat_e: which uses 'even' data to estimate, CV error matrix on 'odd' data.
}
\description{
The hCV method with DP algorithm to find the optimal number of change-points
}
