% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copss.R
\name{hdmean_objfun_cv}
\alias{hdmean_objfun_cv}
\title{High-dimensinal CV objective function - Cross-validation value}
\usage{
hdmean_objfun_cv(Xo, Xe, W, cp_hat)
}
\arguments{
\item{Xo}{dataset used to estimate, matrix of n1*p}

\item{Xe}{dataset used to CV, matrix of n2*p}

\item{W}{weight matrix, p*p}

\item{cp_hat}{The estimated change-point, with 0 and 1 included}
}
\value{
cv_objfun(Xo, Xe, W, cp_hat) - L * p
}
\description{
The high dimensional mean change model, the CV function.
}
\details{
NOTE: Never use this function in hBIC_dp_hausdorff, bacause we need 
  the cv_objfun(X, X, W, cp_hat) in hBIC to calculate the fitting error
  and then plus the penalty, which is calculated by 
  calculate_penalty(X, normalized = TRUE, penaltyform = "hBIC", alph, c0)
  The p has already been included in penalty term, if we use this function
  again, we'll add another p.
}
