% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copss.R
\name{trace_R_square}
\alias{trace_R_square}
\title{Calculate the trace of \eqn{R^2} using R}
\usage{
trace_R_square(X)
}
\arguments{
\item{X}{n*p data matrix}
}
\value{
the value of trace(\eqn{R^2})
}
\description{
Calculate the trace of \eqn{R^2}, \eqn{R^2} is correlation matrix of data X
}
\details{

}
\examples{
X <- matrix(1:100, 20, 5)
trace_R_square(X)

# X <- matrix(rnorm(2048 * 6000), 2048, 6000)
# system.time( tr2 <- trace_R_square_cpp(X) )
# user   system  elapsed 
# 977.470   63.005 1041.387

# system.time( tr2 <- trace_R_square(X) )
# user  system elapsed 
# 350.799 220.544 571.491 
}
