% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{trace_R_square_cpp}
\alias{trace_R_square_cpp}
\title{Calculate the trace of \eqn{R^2} using Rcpp codes}
\usage{
trace_R_square_cpp(X)
}
\arguments{
\item{X}{n*p data matrix}
}
\value{
the value of trace(\eqn{R^2})
}
\description{
Calculate the trace of \eqn{R^2}, \eqn{R^2} is correlation matrix of data X
}
\details{

}
\examples{
X = matrix(1:100, 20, 5)
trace_R_square_cpp(X)

}
